pyproject.toml
[build-system]
requires = [
    "setuptools",
    "versioneer",
]
build-backend = "setuptools.build_meta"

[project]
name = "AcceleratorSimFrame"
description = "A python framework for particle accelerator simulations"
readme = "README.md"
dynamic = ["version"]
authors = [
    {name = "James Jones", email = "james.jones@stfc.ac.uk"},
]
license = "GPL-3.0-or-later"
maintainers = [
    {name = "James Jones", email = "james.jones@stfc.ac.uk"},
    {name = "Alex Brynes", email = "alexander.brynes@stfc.ac.uk"},
    {name = "Mark Johnson", email = "mark.johnson@stfc.ac.uk"},
]
dependencies = [
    "deepdiff>=5",
    "h5py>=2.10",
    "munch>=2.5",
    "numpy>=2",
    "tqdm>=4",
    "PyQt5>=5.1",
    "PyYAML>=5.3",
    "mpl-axes-aligner>=1.1",
    "lox>=0.11",
    "fastKDE>=2.1.5",
    "pydantic>=2.5.3",
    "attrs>=23.2.0",
    "ocelot-desy==25.06.0",
    "scipy>=1.5",
    "soliday.sdds",
    "easygdf>=2.1.1",
    "xsuite>=0.39.0",
    "paramiko",
    "deap",
    "pyqtgraph",
    "numba",
    "pyfftw",
    "numexpr",
    "cheetah-accelerator>=0.7.5",
    "openpmd-beamphysics>=0.10",
    "xopt",
]

requires-python = ">= 3.10"
classifiers = [
    "Programming Language :: Python :: 3.10",
#    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
]

[project.optional-dependencies]

test = [
    "pytest",
    "pytest-cov",
    "MasterLattice",
]

docs = [
    "sphinx>=8",
    "sphinx-rtd-theme",
    "myst-nb",
    "sphinx_autodoc_typehints>=3",
    "MasterLattice",
    "sphinxcontrib-bibtex",
]

[project.urls]
Homepage = "https://github.com/astec-stfc/simframe"
Repository = "https://github.com/astec-stfc/simframe.git"
Issues = "https://github.com/astec-stfc/simframe/-/issues/"

[tool.coverage.run]
omit = [
    # omit this single file
    "_version.py",
    "SimulationFramework/Codes/MAD8/MAD8.py",
    "SimulationFramework/Modules/Beams/mad8.py",
    ]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"mypackage" = ["**/*.yaml"]
